requirements.txt
python_dateutil==2.9.0.post0
Requests==2.32.3
selenium==4.23.0
webdriver_manager==4.0.1
pytest==8.2.1